University FAQ Agent/requirements.txt
openai
langchain
streamlit
python-dotenv
faiss-cpu
pandas